resolves #22 used combine reducers to seperate reducers by category, will continue to use single action file per Jons advice
